diagnostics rework

+++ src/responder.h
@@ -4,7 +4,7 @@
 #include "command.h"
 #include "handler.h"
 
-void monitorData(uint8_t arg1);
+void monitorData(int arg1);
 extern handler Handler;
 bool handleMsg(command cmd);
 void diagReadout(uint8_t arg1);
